fw_cfg: move assert() and linking of fw_cfg device to the machine into instance_init()
In preparation for calling fw_cfg_init1() during realize rather than during
init, move the assert() checking for existing fw_cfg devices and the linking
of the device to the machine with object_property_add_child() to a new
fw_cfg instance_init() function.

The assert() will be removed in a follow-up patch but this ensures that we still
fail hard as before if more than one fw_cfg device is instantiated by accident.

Signed-off-by: Mark Cave-Ayland <mark.cave-ayland@ilande.co.uk>