ppc: use host clock rather than cpu ticks to calculate tb_offset
The existing timebase migration code uses cpu_get_host_ticks() to calculate the
timebase offset between source and destination, however this does not work on
all non-PPC platforms e.g. on Intel host with TCG guest cpu_get_host_ticks()
is mapped to the TSC register and so is independent of the timebase frequency.

Switch to using the host clock to calculate tb_offset which returns the time
difference in ns between source and destination, adjusting by the timebase
frequency as required.

Signed-off-by: Mark Cave-Ayland <mark.cave-ayland@ilande.co.uk>