adb: add status field for holding information about the last ADB request
As inspired by the description of the internal ADB registers in "Macintosh
Technology in the Common Hardware Reference Platform", add a field representing
the status of the last ADB request.

Currently only 2 bits are defined: one to indicate if the request timed out (no
reply) and another to indicate whether the request was the result of an autopoll
operation.

Signed-off-by: Mark Cave-Ayland <mark.cave-ayland@ilande.co.uk>