mac_via: rework ADB state machine to be compatible with both MacOS and Linux
The existing ADB state machine is designed to work with Linux which has a different
interpretation of the state machine detailed in "Guide to the Macintosh Family
Hardware".

Rework the ADB state machine so that it can enumerate and autopoll the ADB under
both Linux and MacOS, including the additional of the new adb_autopoll_block() and
adb_autopoll_unblock() functions.

Signed-off-by: Mark Cave-Ayland <mark.cave-ayland@ilande.co.uk>