mac_via: don't raise interrupt for bus timeout upon switch to IDLE mode
This was required for Linux < 5.6 in order to allow the ADB state machine
to work correctly. Unfortunately with this interrupt enabled MacOS gets
stuck in a loop when issuing its final TALK command after ADB enumeration.

Remove the setting of the interrupt at the expense of not being able to
use ADB on Linux < 5.6.

Signed-off-by: Mark Cave-Ayland <mark.cave-ayland@ilande.co.uk>