mac_via: don't try and optimise timers
The code is wrong here: by calling via1_VBL_update() and via1_one_second_update()
for every VIA1 port B read/write then the timers constantly get reset without
being able to be triggered.

This fixes the Ticks (0x16a) global ensuring that it constantly increases which is
seemingly used for settling delays in the next stage bootloader.

This code was originally introduced as a performance optmisation, however it is
likely that this was a side-effect of the 60Hz timer running too quickly. So let's
just remove it to simplify things.