WIP: esp: defer IRQ until SCSI layer delivers data
MacOS seemingly ignores waiting for DRQ to be asserted before attempting to
perform a "fast" PDMA read. This means that in the case where the SCSI
layer doesn't return immediately, it attempts to read an uninitialised buffer
which returns 0s to the client but also messes our internal state.

Attempt to fix this by delaying the "command complete" IRQ until data from the
device is present in the buffer.

FIXME: this seems to make NextStep unhappy, need to study a bit more.