[RFC] coroutine-sigaltstack: Add SIGUSR2 mutex
Modifying signal handlers or the signal handling stack is a
process-global operation.  When two threads run coroutine-sigaltstack's
qemu_coroutine_new() concurrently, thay may interfere with each other,
e.g.:

- One of the threads may revert the SIGUSR2 handler back to the default
  between the other thread setting up coroutine_trampoline() as the
  handler and raising SIGUSR2.  That SIGUSR2 will then lead to the
  process exiting.

- Both threads may set up their coroutine stack with sigaltstack()
  simultaneously, so that only one of them sticks.  Both then raise
  SIGUSR2, which goes to each of the threads separately, but both signal
  handler invocations will then use the same stack, which cannot work.

We have to ensure that only one thread at a time can modify the
process-global SIGUSR2 handler and the signal handling stack.  To do so,
wrap the whole section where that is done in a mutex.

Signed-off-by: Max Reitz <mreitz@redhat.com>