esp: handle non-DMA transfers from the target one byte at a time
The initial implementation of non-DMA transfers was based upon analysis of traces
from the MacOS toolbox ROM for handling unaligned reads. Further testing from
attempting to boot CDROMs shows that incoming data must be transferred one byte
at a time rather than waiting for the FIFO to be full.

Adjust the non-DMA code accordingly which requires special handling for managing
the change from DATA IN to STATUS phase as this needs to occur when the final byte
is read out from the FIFO, and not at the end of the transfer when it is loaded.

Signed-off-by: Mark Cave-Ayland <mark.cave-ayland@ilande.co.uk>