macio: use address space API to dispatch aliased reads and writes
The memory_region_dispatch_read/write() APIs assume that they are dispatching
to an IO region, and so fault when accessing RAM/ROM regions such as the PROM
in the Sonic network adapter.

Switch to using the address_space_read/write() APIs which will correctly handle
ROM/RAM-based accesses accordingly.