dp8393x: fix PROM checksum and MAC address storage
The checksum used by MacOS to validate the PROM content is an exclusive-OR
rather than a sum over the corresponding bytes. In addition the MAC address
must be stored in bit-reversed format as indicated in comments in Linux's
macsonic.c.

With the PROM contents fixed MacOS starts to probe the device registers
when AppleTalk is enabled in the Control Panel.

TODO: this is a change to the checksum which was originally introduced in
commit 89ae0ff9b7 "net/dp8393x: add PROM to store MAC address" for the MIPS
Jazz. Does MIPS Jazz have a separate PROM configuration without the bit
inversion?

Signed-off-by: Mark Cave-Ayland <mark.cave-ayland@ilande.co.uk>