hw/audio/asc: improve playback buffer management code
The following code in generate_fifo() and generate_wavetable()

    uint8_t *buf = s->mixbuf + s->pos;

and this one in write_audio()

        wbytes = AUD_write(s->voice,
                           s->mixbuf + (pos << (s->shift - 1)),
                           nbytes);

looks highly suspicious. s->pos and pos count in samples (the
code in this file uses samples as a synonym for QEMU frames)
but a sample has two bytes, which means the buffer pointer arith-
metic is wrong.

Instead of trying to gradually fix the code, rewrite the buffer
management code in asc_out_cb() and drop the now unused function
write_audio().

Signed-off-by: Volker Rümelin <vr_qemu@t-online.de>
Message-Id: <20220327185456.12695-2-vr_qemu@t-online.de>